Simplify the version number to just an Integer

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppstier/service/ProtectLevelCalculator.kt
@@ -31,7 +31,7 @@ class ProtectLevelCalculator(
   private val clock: Clock,
   private val communityApiClient: CommunityApiClient,
   private val assessmentApiService: AssessmentApiService,
-  @Value("\${calculation.version}")private val calculationVersion: Float
+  @Value("\${calculation.version}")private val calculationVersion: Int
 ) {
 
   fun calculateProtectLevel(
@@ -115,9 +115,9 @@ class ProtectLevelCalculator(
         val additionalFactorsPoints = getAdditionalFactorsAssessmentComplexityPoints(offenderAssessment)
         val breachRecallPoints = getBreachRecallComplexityPoints(crn, convictions)
 
-        val violenceArsonPoints = if (calculationVersion >= 2.0) getArsonOrViolencePoints(convictions) else 0
+        val violenceArsonPoints = if (calculationVersion >= 2) getArsonOrViolencePoints(convictions) else 0
 
-        val tenMonthsPlusOrIndeterminatePoints = if (calculationVersion >= 2.0) getSentenceLengthPoints(convictions) else 0
+        val tenMonthsPlusOrIndeterminatePoints = if (calculationVersion >= 2) getSentenceLengthPoints(convictions) else 0
 
         additionalFactorsPoints + breachRecallPoints + violenceArsonPoints + tenMonthsPlusOrIndeterminatePoints
       }
